Tests and shared functions sorted
