+++ test/conformance/api/generatename_test.go
@@ -95,11 +95,11 @@ func canServeRequests(t *testing.T, clients *test.Clients, route *v1alpha1.Route
 		clients.KubeClient,
 		t.Logf,
 		domain,
-		pkgTest.Retrying(pkgTest.MatchesAllOf(pkgTest.IsStatusOK, pkgTest.MatchesBody(test.HelloWorldText)), http.StatusNotFound),
+		pkgTest.Retrying(pkgTest.MatchesAllOf(pkgTest.IsStatusOK, pkgTest.MatchesBody(helloWorldText)), http.StatusNotFound),
 		"WaitForEndpointToServeText",
 		test.ServingFlags.ResolvableDomain)
 	if err != nil {
-		return fmt.Errorf("the endpoint for Route %s at domain %s didn't serve the expected text %q: %v", route.Name, domain, test.HelloWorldText, err)
+		return fmt.Errorf("the endpoint for Route %s at domain %s didn't serve the expected text %q: %v", route.Name, domain, helloWorldText, err)
 	}
 
 	return nil
@@ -114,7 +114,7 @@ func TestServiceGenerateName(t *testing.T) {
 
 	generateName := generateNamePrefix(t)
 	names := test.ResourceNames{
-		Image: test.Helloworld,
+		Image: helloworld,
 	}
 
 	// Cleanup on test failure.
@@ -151,7 +151,7 @@ func TestRouteAndConfigGenerateName(t *testing.T) {
 
 	generateName := generateNamePrefix(t)
 	names := test.ResourceNames{
-		Image: test.Helloworld,
+		Image: helloworld,
 	}
 
 	test.CleanupOnInterrupt(func() { test.TearDown(clients, names) })
