+++ test/conformance/api/single_threaded_test.go
@@ -39,7 +39,7 @@ func TestSingleConcurrency(t *testing.T) {
 
 	names := test.ResourceNames{
 		Service: test.ObjectNameForTest(t),
-		Image:   test.SingleThreadedImage,
+		Image:   singleThreadedImage,
 	}
 	test.CleanupOnInterrupt(func() { test.TearDown(clients, names) })
 	defer test.TearDown(clients, names)
