+++ test/conformance/runtime/cgroup_test.go
@@ -63,7 +63,7 @@ func TestMustHaveCgroupConfigured(t *testing.T) {
 		"/sys/fs/cgroup/cpu/cpu.cfs_quota_us":         cpuLimit * 1000 * 100, // 1000 millicore * 100
 		"/sys/fs/cgroup/cpu/cpu.shares":               cpuRequest * 1024}     // CPURequests * 1024
 
-	_, ri, err := test.FetchRuntimeInfo(t, clients, &test.Options{ContainerResources: resources})
+	_, ri, err := fetchRuntimeInfo(t, clients, &test.Options{ContainerResources: resources})
 	if err != nil {
 		t.Fatalf("Error fetching runtime info: %v", err)
 	}
@@ -91,7 +91,7 @@ func TestMustHaveCgroupConfigured(t *testing.T) {
 func TestShouldHaveCgroupReadOnly(t *testing.T) {
 	t.Parallel()
 	clients := test.Setup(t)
-	_, ri, err := test.FetchRuntimeInfo(t, clients, &test.Options{})
+	_, ri, err := fetchRuntimeInfo(t, clients, &test.Options{})
 	if err != nil {
 		t.Fatalf("Error fetching runtime info: %v", err)
 	}
