+++ test/conformance/runtime/envvars_test.go
@@ -32,7 +32,7 @@ import (
 func TestShouldEnvVars(t *testing.T) {
 	t.Parallel()
 	clients := test.Setup(t)
-	names, ri, err := test.FetchRuntimeInfo(t, clients, &test.Options{})
+	names, ri, err := fetchRuntimeInfo(t, clients, &test.Options{})
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -61,7 +61,7 @@ func TestMustEnvVars(t *testing.T) {
 	if err != nil {
 		t.Fatal("Invalid PORT value in MustEnvVars")
 	}
-	_, ri, err := test.FetchRuntimeInfo(t, clients, &test.Options{
+	_, ri, err := fetchRuntimeInfo(t, clients, &test.Options{
 		ContainerPorts: []corev1.ContainerPort{
 			{ContainerPort: int32(port)},
 		},
