+++ test/conformance/runtime/file_descriptor_test.go
@@ -29,7 +29,7 @@ import (
 func TestShouldHaveStdinEOF(t *testing.T) {
 	clients := test.Setup(t)
 
-	_, ri, err := test.FetchRuntimeInfo(t, clients, &test.Options{})
+	_, ri, err := fetchRuntimeInfo(t, clients, &test.Options{})
 	if err != nil {
 		t.Fatalf("Error fetching runtime info: %v", err)
 	}
